ingestion/requirements.txt
functions-framework==3.*
google-cloud-storage==2.*
google-cloud-bigquery==3.*
google-cloud-aiplatform==1.*
google-auth==2.*
chromadb==1.5.0
pyyaml==6.*
